Fixed bug in the sorting of PairForces, energy conservation confirmed with VelVerlet.
Not resetting the texture reference to particleIndex when switching the arrays after the cub radix sorting caused that the particleIndex didnt correspond to the current cellIndex one every two steps. Causing total disaster.
The infraestructure to measure the energy was added to TwoStepVelVerlet and PairForces, encoding it in force.w