DPD is now working. Fixed crash when using Matrix.
Big changes to DPD module, now it maintains the desired temperature.
Improvements to the Vector and Matrix containers, now they are faster and safer.

The destructor issue in the global Vector objects is still a problem, although it is only a problem when the program exits, so it is only annoying.
The current solution has no memory leaks, so for the moment Ill leave it like that.